requirements.txt
pdfrw
reportlab
PyPDF2
GitPython
pandas